Refactorisation du système de traduction. On intègre les traduction des favoris aux trad global
